algunos cambios de nombres y en la arquitectura. cli_example
